uint_8t

+++ src/arch_riscv.h
@@ -71,10 +71,10 @@ typedef enum {
 
 #define FUNCHOOK_ENTRY_AT_PAGE_BOUNDARY 1
 
-typedef uint32_t insn_t;
+typedef uint8_t insn_t;
 
 typedef struct funchook_entry {
-    uint32_t transit[TRANSIT_CODE_SIZE];
+    uint8_t transit[TRANSIT_CODE_SIZE];
     void *original_target_func;
     void *target_func;
     void *hook_func;
@@ -82,9 +82,9 @@ typedef struct funchook_entry {
     void *user_data;
     char *arg_types;
     uint32_t flags;
-    uint32_t trampoline[TRAMPOLINE_SIZE];
-    uint32_t old_code[JUMP32_SIZE];
-    uint32_t new_code[JUMP32_SIZE];
+    uint8_t trampoline[TRAMPOLINE_SIZE];
+    uint8_t old_code[JUMP32_SIZE];
+    uint8_t new_code[JUMP32_SIZE];
 } funchook_entry_t;
 
 typedef struct {
